+++ tests/IntegrationTests/DI/InjectionTest.php
@@ -22,6 +22,7 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
     const DEFINITION_REFLECTION = 1;
     const DEFINITION_ANNOTATIONS = 2;
     const DEFINITION_ARRAY = 3;
+    const DEFINITION_PHP = 4;
 
     /**
      * PHPUnit data provider: generates container configurations for running the same tests for each configuration possible
@@ -35,11 +36,11 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         $containerReflection->useAnnotations(false);
         $containerReflection->addDefinitions(
             array(
-                'foo' => 'bar',
+                'foo'                                     => 'bar',
                 'IntegrationTests\DI\Fixtures\Interface1' => array(
                     'class' => 'IntegrationTests\DI\Fixtures\Implementation1',
                 ),
-                'namedDependency' => array(
+                'namedDependency'                         => array(
                     'class' => 'IntegrationTests\DI\Fixtures\Class2',
                 ),
             )
@@ -51,11 +52,11 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         $containerAnnotations->useAnnotations(true);
         $containerAnnotations->addDefinitions(
             array(
-                'foo' => 'bar',
+                'foo'                                     => 'bar',
                 'IntegrationTests\DI\Fixtures\Interface1' => array(
                     'class' => 'IntegrationTests\DI\Fixtures\Implementation1',
                 ),
-                'namedDependency' => array(
+                'namedDependency'                         => array(
                     'class' => 'IntegrationTests\DI\Fixtures\Class2',
                 ),
             )
@@ -67,7 +68,7 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
         $containerArray->useAnnotations(false);
         $containerArray->addDefinitions(
             array(
-                'foo' => 'bar',
+                'foo'                                          => 'bar',
                 'IntegrationTests\DI\Fixtures\Class1'          => array(
                     'scope'       => Scope::PROTOTYPE(),
                     'properties'  => array(
@@ -78,7 +79,7 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
                     ),
                     'constructor' => array(
                         'param1' => 'IntegrationTests\DI\Fixtures\Class2',
-                        'param2' => 'IntegrationTests\DI\Fixtures\Interface1'
+                        'param2' => 'IntegrationTests\DI\Fixtures\Interface1',
                     ),
                     'methods'     => array(
                         'method1' => 'IntegrationTests\DI\Fixtures\Class2',
@@ -95,16 +96,45 @@ class InjectionTest extends \PHPUnit_Framework_TestCase
                     'class' => 'IntegrationTests\DI\Fixtures\Implementation1',
                     'scope' => 'singleton',
                 ),
-                'namedDependency' => array(
+                'namedDependency'                              => array(
                     'class' => 'IntegrationTests\DI\Fixtures\Class2',
                 ),
             )
         );
 
+        // Test with a container using PHP configuration
+        $containerPHP = new Container();
+        $containerPHP->useReflection(false);
+        $containerPHP->useAnnotations(false);
+        $containerPHP->set('foo', 'bar');
+        $containerPHP->set('IntegrationTests\DI\Fixtures\Class1')
+            ->withScope(Scope::PROTOTYPE())
+            ->withProperty('property1', 'IntegrationTests\DI\Fixtures\Class2')
+            ->withProperty('property2', 'IntegrationTests\DI\Fixtures\Interface1')
+            ->withProperty('property3', 'namedDependency')
+            ->withProperty('property4', 'foo')
+            ->withConstructor(
+                array(
+                    'param1' => 'IntegrationTests\DI\Fixtures\Class2',
+                    'param2' => 'IntegrationTests\DI\Fixtures\Interface1',
+                )
+            )
+            ->withMethod('method1', array('IntegrationTests\DI\Fixtures\Class2'))
+            ->withMethod('method2', array('IntegrationTests\DI\Fixtures\Interface1'))
+            ->withMethod('method3', array('param1' => 'namedDependency', 'param2' => 'foo'));
+        $containerPHP->set('IntegrationTests\DI\Fixtures\Class2');
+        $containerPHP->set('IntegrationTests\DI\Fixtures\Implementation1');
+        $containerPHP->set('IntegrationTests\DI\Fixtures\Interface1')
+            ->bindTo('IntegrationTests\DI\Fixtures\Implementation1')
+            ->withScope(Scope::SINGLETON());
+        $containerPHP->set('namedDependency')
+            ->bindTo('IntegrationTests\DI\Fixtures\Class2');
+
         return array(
             array(self::DEFINITION_REFLECTION, $containerReflection),
             array(self::DEFINITION_ANNOTATIONS, $containerAnnotations),
             array(self::DEFINITION_ARRAY, $containerArray),
+            array(self::DEFINITION_PHP, $containerPHP),
         );
     }